Add files via upload
So on 18 July, the project was taken forward by simply loading the data on the file and training the model just to see the losses the model delivers. In next 2 days the model has to be optimized to make the model to predict output more precisely along with the functions to make it easy to understand and interpret.